+++ packages/catalog-protocol/src/index.ts
@@ -1,5 +1,6 @@
 export * from './canonicalize'
 export * from './contribution-auth'
+export * from './contract'
 export * from './hash'
 export * from './normalization'
 export * from './projector'